+++ R/colwise-arrange.R
@@ -29,31 +29,45 @@
 #' arrange_all(df, desc)
 #' # ->
 #' arrange(df, across(everything(), desc))
-arrange_all <- function(.tbl, .funs = list(), ..., .by_group = FALSE) {
+arrange_all <- function(.tbl,
+                        .funs = list(),
+                        ...,
+                        .by_group = FALSE,
+                        .locale = dplyr_locale()) {
   lifecycle::signal_stage("superseded", "arrange_all()")
   funs <- manip_all(.tbl, .funs, enquo(.funs), caller_env(), .include_group_vars = TRUE, ..., .caller = "arrange_all")
   if (!length(funs)) {
     funs <- syms(tbl_vars(.tbl))
   }
-  arrange(.tbl, !!!funs, .by_group = .by_group)
+  arrange(.tbl, !!!funs, .by_group = .by_group, .locale = .locale)
 }
 #' @rdname arrange_all
 #' @export
-arrange_at <- function(.tbl, .vars, .funs = list(), ..., .by_group = FALSE) {
+arrange_at <- function(.tbl,
+                       .vars,
+                       .funs = list(),
+                       ...,
+                       .by_group = FALSE,
+                       .locale = dplyr_locale()) {
   lifecycle::signal_stage("superseded", "arrange_at()")
   funs <- manip_at(.tbl, .vars, .funs, enquo(.funs), caller_env(), .include_group_vars = TRUE, ..., .caller = "arrange_at")
   if (!length(funs)) {
     funs <- tbl_at_syms(.tbl, .vars, .include_group_vars = TRUE)
   }
-  arrange(.tbl, !!!funs, .by_group = .by_group)
+  arrange(.tbl, !!!funs, .by_group = .by_group, .locale = .locale)
 }
 #' @rdname arrange_all
 #' @export
-arrange_if <- function(.tbl, .predicate, .funs = list(), ..., .by_group = FALSE) {
+arrange_if <- function(.tbl,
+                       .predicate,
+                       .funs = list(),
+                       ...,
+                       .by_group = FALSE,
+                       .locale = dplyr_locale()) {
   lifecycle::signal_stage("superseded", "arrange_if()")
   funs <- manip_if(.tbl, .predicate, .funs, enquo(.funs), caller_env(), .include_group_vars = TRUE, ..., .caller = "arrange_if")
   if (!length(funs)) {
     funs <- tbl_if_syms(.tbl, .predicate, .include_group_vars = TRUE)
   }
-  arrange(.tbl, !!!funs, .by_group = .by_group)
+  arrange(.tbl, !!!funs, .by_group = .by_group, .locale = .locale)
 }