+++ R/grouped-df.r
@@ -298,7 +298,7 @@ vec_split_id_order <- function(x) {
   split_id <- vec_group_loc(x)
   split_id$loc <- new_list_of(split_id$loc, ptype = integer())
 
-  vec_slice(split_id, vec_order(split_id$key))
+  vec_slice(split_id, vec_order_base(split_id$key))
 }
 
 group_intersect <- function(x, new) {