+++ tests/testthat/test-arrange.r
@@ -82,14 +82,81 @@ test_that("arrange handles S4 classes (#1105)", {
   expect_equal(arrange(df, y), df[3:1, ])
 })
 
-test_that("arrange respects locale (#1280)", {
-  df2 <- tibble(words = c("casa", "\u00e1rbol", "zona", "\u00f3rgano"))
+# locale --------------------------------------------------------------
 
-  res <- df2 %>% arrange(words)
-  expect_equal(res$words, sort(df2$words))
+test_that("arrange defaults to the C locale", {
+  x <- c("A", "a", "b", "B")
+  df <- tibble(x = x)
 
-  res <- df2 %>% arrange(desc(words))
-  expect_equal(res$words, sort(df2$words, decreasing = TRUE))
+  res <- arrange(df, x)
+  expect_identical(res$x, c("A", "B", "a", "b"))
+
+  res <- arrange(df, desc(x))
+  expect_identical(res$x, rev(c("A", "B", "a", "b")))
+})
+
+test_that("locale can be set to an English locale", {
+  skip_if_not_installed("stringi", "1.5.3")
+
+  x <- c("A", "a", "b", "B")
+  df <- tibble(x = x)
+
+  res <- arrange(df, x, .locale = "en")
+  expect_identical(res$x, c("a", "A", "b", "B"))
+})
+
+test_that("non-English locales can be used", {
+  skip_if_not_installed("stringi", "1.5.3")
+
+  # Danish `o` with `/` through it sorts after `z` in Danish locale
+  x <- c("o", "\u00F8", "p", "z")
+  df <- tibble(x = x)
+
+  # American English locale puts it right after `o`
+  res <- arrange(df, x, .locale = "en")
+  expect_identical(res$x, x)
+
+  res <- arrange(df, x, .locale = "da")
+  expect_identical(res$x, x[c(1, 3, 4, 2)])
+})
+
+test_that("arrange respects the `dplyr.locale` global option", {
+  skip_if_not_installed("stringi", "1.5.3")
+
+  local_options(dplyr.locale = "da")
+
+  x <- c("o", "\u00F8", "p", "z")
+  df <- tibble(x = x)
+
+  res <- arrange(df, x)
+  expect_identical(res$x, x[c(1, 3, 4, 2)])
+})
+
+test_that("arrange errors if stringi is not installed and a locale identifier is used", {
+  expect_snapshot(error = TRUE, {
+    locale_to_chr_proxy_collate("fr", has_stringi = FALSE)
+  })
+})
+
+test_that("arrange validates `.locale`", {
+  df <- tibble()
+
+  expect_snapshot(error = TRUE, {
+    arrange(df, .locale = 1)
+  })
+  expect_snapshot(error = TRUE, {
+    arrange(df, .locale = c("en_US", "fr_BF"))
+  })
+})
+
+test_that("arrange validates that `.locale` must be one from stringi", {
+  skip_if_not_installed("stringi", "1.5.3")
+
+  df <- tibble()
+
+  expect_snapshot(error = TRUE, {
+    arrange(df, .locale = "x")
+  })
 })
 
 # data ----------------------------------------------------------------